Feature/cleanup

+++ src/main/java/com/brickroad/starcreator_webservice/repository/PlanetRepository.java
@@ -3,6 +3,6 @@ package com.brickroad.starcreator_webservice.repository;
 import com.brickroad.starcreator_webservice.entity.ud.Planet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanetRepository extends JpaRepository<Planet, Integer> {
+public interface PlanetRepository extends JpaRepository<Planet, Long> {
 
 }
