+++ src/main/java/com/brickroad/starcreator_webservice/repository/SectorRepository.java
@@ -3,7 +3,7 @@ package com.brickroad.starcreator_webservice.repository;
 import com.brickroad.starcreator_webservice.entity.ud.Sector;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectorRepository extends JpaRepository<Sector,Integer> {
+public interface SectorRepository extends JpaRepository<Sector, Long> {
