+++ src/main/java/com/brickroad/starcreator_webservice/repository/StarSystemRepository.java
@@ -3,7 +3,7 @@ package com.brickroad.starcreator_webservice.repository;
 import com.brickroad.starcreator_webservice.entity.ud.StarSystem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StarSystemRepository extends JpaRepository<StarSystem, Integer> {
+public interface StarSystemRepository extends JpaRepository<StarSystem, Long> {
